Code is broken for non-linear curves
Test currently holds a version closer to what the final version will need to be. Basically the position function and the integral are disagreeing some where. This is being tested with a solenoid. Currently the solenoid produces large vectors down the centeral axis that spin with the coiling. My hunch is that I may have done some linear algebra wrong, but I can't find the mistake and my brain is going numb. For the 1d wire case the program works well.