Created new doc for rind sim, added particle dynamics (potential issue)
Added a seperate file for rings as I felt that it was annoying to constantly go back and forth (maybe I will add an initial config file). I added particle dynamics and noticed that the field does not seem to create closed motion which concerns me. I think that I should observe cyclic motion of the particle, but I do not see this instead things seem to shoot off to infinity, I should look into this. Attemped to add a color heat map to the vector field, but I consistently ran into a problem with the matplotlib GUI, the vectors became large squares. This can be fixed soon, but finals must be finished first. I read about Faraday's and it seems possible. I will need a way of taking surfaces converting to grids and then computing a "mesh" which I can do linear magnetic flux approximations. This, of course will be really computationally expensive and will most likely need to be moved to c++ and use the GPU inorder to have decent load times.